% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axlim.R
\name{axlim}
\alias{axlim}
\title{Set rounded axis limits}
\usage{
axlim(x, level)
}
\arguments{
\item{x}{A numeric vector of data}

\item{level}{The "level" to round the lower and upper limits to}
}
\value{
A numeric vector with the lower and upper limit
}
\description{
Set rounded axis limits given data and a specified rounding level.
}
\examples{
x <- seq(3,95,1)
axlim(x, 5)
axlim(x, 10)
axlim(x, 30)
}
