% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceiling1.R
\name{ceiling1}
\alias{ceiling1}
\title{Round to ceiling of specified level}
\usage{
ceiling1(x, level)
}
\arguments{
\item{x}{A number}
}
\value{
A rounded number
}
\description{
Raises a number to the nearest 0.1, 0.5, 1, 5, 10, etc.
}
\examples{
# Raise 0.34 to nearest 0.5
ceiling1(0.34, level=0.5)
#' Raise 35.6 to nearest 60
ceiling1(35.6, level=50)
}
