% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_species.R
\name{check_species}
\alias{check_species}
\title{Check marine fish and invertebrate scientific names}
\usage{
check_species(species)
}
\arguments{
\item{species}{A character vector of species scientific names}
}
\value{
A character vector of species scientific names not in FishBase
}
\description{
Checks marine fish and invertebrate scientific names against species in FishBase.
}
\examples{
# Check species names (Gadus morhoa is incorrect)
species <- c("Gadus morhua", "Gadus morhoa")
check_species(species)
}
