% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpal.R
\name{colorpal}
\alias{colorpal}
\title{Create expanded color palette}
\usage{
colorpal(pal, n)
}
\arguments{
\item{pal}{A vector of colors}

\item{n}{Number of colors to be interpolated}
}
\value{
An expanded vector of colors
}
\description{
Creates an expanded color palette from an existing color palette by interpolating new colors.
}
\examples{
# Make 15 color palette from 11 color red-blue palette
library(RColorBrewer)
colorpal(brewer.pal(11, "RdBu"), 15)
}
