% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2brick.R
\name{df2brick}
\alias{df2brick}
\title{Convert a XYZ-layer data frame to a raster brick}
\usage{
df2brick(df, x_col, y_col, z_col, layer_col)
}
\arguments{
\item{df}{A dataframe with XYZ info for many layers}

\item{x_col}{Name of the column with the x info}

\item{y_col}{Name of the column with the y info}

\item{z_col}{Name of the column with the z info}

\item{layer_col}{Name of the column with the layer info}
}
\value{
A raster brick
}
\description{
Converts a XYZ-layer data frame to a raster brick.
}
\examples{
nlayers <- 10
df <- expand.grid(layer=1:nlayers, x=1:10, y=1:10) \%>\%
  arrange(layer) \%>\%
  mutate(z=runif(n=n()))
df_brick <- df2brick(df=df, x_col="x", y_col="y", z_col="z", layer_col="layer")
raster::plot(df_brick)
}
