% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyplot.R, R/plot_empty.R
\name{emptyplot}
\alias{emptyplot}
\alias{emptyplot}
\title{Creates an empty plot}
\usage{
emptyplot()

emptyplot()
}
\description{
Creates an empty plot with both axis extending from 0 to 10.

Creates an empty plot with both axis extending from 0 to 10.
}
\examples{
emptyplot()
emptyplot()
}
