% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_comm_name.R
\name{fb_comm_name}
\alias{fb_comm_name}
\title{Get common name for marine fish and invertebrates}
\usage{
fb_comm_name(species)
}
\arguments{
\item{species}{A character vector of species scientific names}
}
\value{
A dataframe of species common names and their sources
}
\description{
Gets ONE common name for marine fish and invertebrate scientific names. The common name comes from FishBase species table, FishBase English common names,
FishBase non-English common names.
}
\examples{
# Get common names
species <- c("Gadus morhua", "Clupea harengus", "Callinectes sapidus")
fb_comm_name(species)
}
