% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_wday_in_month.R
\name{first_wday_in_month}
\alias{first_wday_in_month}
\title{Find the first X day in month Y in one or multiple years}
\usage{
first_wday_in_month(day, month, years)
}
\arguments{
\item{day}{A month of the year (e.g., "November", "December", etc.)}

\item{years}{A vector of years (e.g., 2014, 2014:2016)}
}
\value{
A vector of dates
}
\description{
Find, for example, the first Saturday in November every year from 2014-2018.
}
\examples{
# Find the first Saturday of November in 2014-2018
first_sats <- first_wday_in_month(day="Saturday", month="November", years=2014:2018)
first_sats
}
