% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floor1.R
\name{floor1}
\alias{floor1}
\title{Round to floor of specified level}
\usage{
floor1(x, level)
}
\arguments{
\item{x}{A number}
}
\value{
A rounded number
}
\description{
Lowers a number to the nearest 0.1, 0.5, 1, 5, 10, etc.
}
\examples{
# Lower 0.34 to nearest 0.25
floor1(0.34, level=0.25)
#' Lower 35.6 to nearest 30
floor1(35.6, level=30)
}
