% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwords.R
\name{nwords}
\alias{nwords}
\title{Number of words}
\usage{
nwords(x)
}
\value{
Number of words in a string
}
\description{
Counts and returns the number of words in each element of a character vector.
}
\examples{
# Number of words
phrases <- c("Words", "Two words", "This is four words.")
nwords(phrases)
}
