% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.R
\name{pval}
\alias{pval}
\title{Extract p-value of linear regression}
\usage{
pval(lmfit)
}
\arguments{
\item{lmfit}{A linear model fit using lm()}
}
\value{
p-value of the linear model fit
}
\description{
Extracts the p-value of a linear regression.
}
