% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\title{Extract r2 of linear regression}
\usage{
r2(lmfit)
}
\arguments{
\item{lmfit}{A linear model fit using lm()}
}
\value{
r2 of the linear model fit
}
\description{
Extracts the r2 of a linear regression.
}
