% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds2csv.R
\name{rds2csv}
\alias{rds2csv}
\title{Convert RDS file to CSV file}
\usage{
rds2csv(file)
}
\arguments{
\item{file}{RDS file name including path}
}
\value{
A CSV file in the same folder and with the same name as the RDS file
}
\description{
Converts an RDS file containing an R dataframe to a CSV.
}
