% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilience.R
\name{resilience}
\alias{resilience}
\title{Download resilience values from FishBase/SeaLifeBase}
\usage{
resilience(species)
}
\arguments{
\item{species}{A character vector of species scientific names to look up}
}
\value{
A dataframe of resilience values from FishBase/SeaLifeBase
}
\description{
Downloads resilience values from FishBase/SeaLifeBase.
}
\examples{
# Download cleaned FishBase life history data
species <- c("Callinectes sapidus", "Gadus morhua")
resilience(species=species)
}
