% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundf.R
\name{roundf}
\alias{roundf}
\title{Round and format to specified decimal place}
\usage{
roundf(x, decimals)
}
\arguments{
\item{x}{Number(s)}

\item{decimals}{Number of decimals}
}
\value{
Formatted number(s) in character format
}
\description{
Rounds and formats a number to specified decimal place.
}
\examples{
roundf(0.345957, 2)
}
