% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentcase.R
\name{sentcase}
\alias{sentcase}
\title{Change text to sentence case}
\usage{
sentcase(x)
}
\arguments{
\item{x}{A character vector in any case}
}
\value{
A character vector in sentence case
}
\description{
Changes string vector to sentence case.
}
\examples{
# Convert to sentence case
sentcase(x=c("Gadus Morhua", "gadus morhua", "GADUS morhua"))
}
