% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R, R/slope.R
\name{slope}
\alias{slope}
\alias{slope}
\title{Extract r2 of linear regression}
\usage{
slope(lmfit)

slope(lmfit)
}
\arguments{
\item{lmfit}{A linear model fit using lm()}

\item{lmfit}{A linear model fit using lm()}
}
\value{
r2 of the linear model fit

Slope of the linear model fit
}
\description{
Extracts the r2 of a linear regression.

Extracts the slope of a linear regression.
}
