% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_names.R
\name{suggest_names}
\alias{suggest_names}
\title{Suggest correct names for invalid scientific names}
\usage{
suggest_names(species)
}
\arguments{
\item{species}{A character vector of species scientific names}
}
\value{
A list of suggestions of correct scientific names for invalid scientific names.
}
\description{
Suggests correct names for invalid scientific names of marine fish and invertebrates.
The function first suggests valid synonyms on FishBase and SeaLifeBase. If no valid
synonyms are available, the function suggests potential names by fuzzy matching
the available FishBase/SeaLifeBase names. The function is essentially an extension
of rfishbase::validate_names() but produces a more organized and thorough output.
}
\examples{
# Suggest scientific names for invalid scientific names
species <- c("Neoplatycephalus richardsoni", "Tetrapturus albidus", "Gadus morhuaa", "Loligo pealeii")
suggest_names(species)
}
