% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{taxa}
\alias{taxa}
\title{Get taxonomic info for marine fish and invertebrates}
\usage{
taxa(species)
}
\arguments{
\item{species}{A character vector of species scientific names}
}
\value{
A dataframe of taxonomic information
}
\description{
Get taxonomic info for marine fish and invertebrates from FishBase/SeaLifeBase.
}
\examples{
# Get taxonomic info
species <- c("Gadus morhua", "Centropristis striata")
taxa(species)
}
