% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcolor.R
\name{tcolor}
\alias{tcolor}
\title{Make transparent color}
\usage{
tcolor(color, alpha)
}
\arguments{
\item{color}{A color (e.g., "green", "#4FFF33)}

\item{alpha}{The transparency level (0-1; 0=invisible, 1=opaque)}
}
\value{
A color with transparency
}
\description{
Makes a color represented by a word ("green") or hex code ("#4FFF33) transparent.
}
\examples{
# Plot using transparent red
plot(1:10, 1:10, col=tcolor("red", alpha=0.4))
}
