% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniq.R
\name{uniq}
\alias{uniq}
\title{Returns a list of sorted unique values}
\usage{
uniq(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A sorted and unique character vector
}
\description{
Returns a list of sorted unique values.
}
