% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_duplicated.R
\name{which_duplicated}
\alias{which_duplicated}
\title{Identify duplicated items}
\usage{
which_duplicated(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector of unique duplicated items
}
\description{
Returns vector of duplicated items.
}
\examples{
# Get taxonomic info
vec <- c("Gadus morhua", "Gadus morhua", "Centropristis striata")
which_duplicated(vec)
}
