% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{complete}
\alias{complete}
\title{Check completeness of dataframe}
\usage{
complete(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A vector showing the the number of NA values in a dataframe
}
\description{
Checks the completeness of columns in a dataframe.
}
