% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishbase.R
\name{fishbase}
\alias{fishbase}
\title{Download cleaned FishBase/SeaLifeBase life history data}
\usage{
fishbase(dataset, species, level = "species", cleaned = F, add_taxa = T)
}
\arguments{
\item{dataset}{FishBase/SeaLifeBase dataset to download: species, stocks, lw, vonb, ecology, maturity, fecundity, reproduction, morphology, ecosystem, speed}

\item{species}{A character vector of species scientific names to look up}

\item{level}{Download life history data for just the provided species ("species") or for all species in the genera ("genus"), families ("family"), or orders ("order") represented in the requested species list.}

\item{cleaned}{FALSE means you get all of the data and TRUE means you get a cleaned subset of important columns}

\item{add_taxa}{TRUE means taxonomic information is added to the life history data}
}
\value{
A dataframe of life history traits from FishBase/SeaLifeBase
}
\description{
Downloads FishBase and SeaLifeBase life history data using rfishbase. The download comes cleaned and the function includes the option to return life history for all of the species included in either the genera or the families of the species requested (this is useful when calculating genus- or family-level life history averages.)
}
\examples{
# Download cleaned FishBase life history data
species <- c("Callinectes sapidus", "Gadus morhua")
lh_data <- fishbase(dataset="species", species=species, level="species", cleaned=T)
}
