% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope}
\alias{slope}
\title{Extract slope of linear regression}
\usage{
slope(lmfit)
}
\arguments{
\item{lmfit}{A linear model fit using lm()}
}
\value{
Slope of the linear model fit
}
\description{
Extracts the slope of a linear regression.
}
